+++ src/lib/components/DatePicker.react.js
@@ -45,6 +45,9 @@ const DatePicker = (props) => {
             "initialMonth",
             "class_name",
             "value",
+            "persistence",
+            "persisted_props",
+            "persistence_type",
         ],
         props
     );
@@ -72,8 +75,8 @@ const DatePicker = (props) => {
 DatePicker.displayName = "DatePicker";
 
 DatePicker.defaultProps = {
-    persisted_props: ['value'],
-    persistence_type: 'local',
+    persisted_props: ["value"],
+    persistence_type: "local",
 };
 
 DatePicker.propTypes = {
@@ -245,20 +248,20 @@ DatePicker.propTypes = {
         PropTypes.string,
         PropTypes.number,
     ]),
-        
+
     /**
      * Properties whose user interactions will persist after refreshing the
-     * component or the page. 
+     * component or the page.
      */
-    persisted_props: PropTypes.arrayOf(PropTypes.oneOf(['value'])),
-        
+    persisted_props: PropTypes.arrayOf(PropTypes.oneOf(["value"])),
+
     /**
      * Where persisted user changes will be stored:
      * memory: only kept in memory, reset on page refresh.
      * local: window.localStorage, data is kept after the browser quit.
      * session: window.sessionStorage, data is cleared once the browser quit.
      */
-    persistence_type: PropTypes.oneOf(['local', 'session', 'memory']),
+    persistence_type: PropTypes.oneOf(["local", "session", "memory"]),
 
     /**
      * Placeholder, displayed when date is not selected