+++ src/lib/components/DateRangePicker.react.js
@@ -60,6 +60,9 @@ const DateRangePicker = (props) => {
             "maxDate",
             "initialMonth",
             "class_name",
+            "persistence",
+            "persisted_props",
+            "persistence_type",
         ],
         props
     );
@@ -88,8 +91,8 @@ const DateRangePicker = (props) => {
 DateRangePicker.displayName = "DateRangePicker";
 
 DateRangePicker.defaultProps = {
-    persisted_props: ['value'],
-    persistence_type: 'local',
+    persisted_props: ["value"],
+    persistence_type: "local",
 };
 
 DateRangePicker.propTypes = {
@@ -261,21 +264,21 @@ DateRangePicker.propTypes = {
         PropTypes.string,
         PropTypes.number,
     ]),
-    
+
     /**
      * Properties whose user interactions will persist after refreshing the
-     * component or the page. 
+     * component or the page.
      */
-    persisted_props: PropTypes.arrayOf(PropTypes.oneOf(['value'])),
-    
+    persisted_props: PropTypes.arrayOf(PropTypes.oneOf(["value"])),
+
     /**
      * Where persisted user changes will be stored:
      * memory: only kept in memory, reset on page refresh.
      * local: window.localStorage, data is kept after the browser quit.
      * session: window.sessionStorage, data is cleared once the browser quit.
      */
-    persistence_type: PropTypes.oneOf(['local', 'session', 'memory']),
-    
+    persistence_type: PropTypes.oneOf(["local", "session", "memory"]),
+
     /**
      * Placeholder, displayed when date is not selected
      */