+++ src/lib/components/MultiSelect.react.js
@@ -9,7 +9,17 @@ import { renderDashComponents } from "dash-extensions-js";
  */
 const MultiSelect = (props) => {
     const { setProps, data, class_name } = props;
-    let nProps = omit(["setProps", "data", "class_name"], props);
+    let nProps = omit(
+        [
+            "setProps",
+            "data",
+            "class_name",
+            "persistence",
+            "persisted_props",
+            "persistence_type",
+        ],
+        props
+    );
     nProps = renderDashComponents(nProps, [
         "label",
         "description",
@@ -36,8 +46,8 @@ MultiSelect.displayName = "MultiSelect";
 
 MultiSelect.defaultProps = {
     data: [],
-    persisted_props: ['value'],
-    persistence_type: 'local',
+    persisted_props: ["value"],
+    persistence_type: "local",
 };
 
 MultiSelect.propTypes = {
@@ -171,20 +181,20 @@ MultiSelect.propTypes = {
         PropTypes.string,
         PropTypes.number,
     ]),
-    
+
     /**
      * Properties whose user interactions will persist after refreshing the
-     * component or the page. 
+     * component or the page.
      */
-    persisted_props: PropTypes.arrayOf(PropTypes.oneOf(['value'])),
-    
+    persisted_props: PropTypes.arrayOf(PropTypes.oneOf(["value"])),
+
     /**
      * Where persisted user changes will be stored:
      * memory: only kept in memory, reset on page refresh.
      * local: window.localStorage, data is kept after the browser quit.
      * session: window.sessionStorage, data is cleared once the browser quit.
      */
-    persistence_type: PropTypes.oneOf(['local', 'session', 'memory']),
+    persistence_type: PropTypes.oneOf(["local", "session", "memory"]),
 
     /**
      * Placeholder, displayed when date is not selected