+++ src/lib/components/RadioGroup.react.js
@@ -16,7 +16,17 @@ const RadioGroup = (props) => {
     return (
         <MantineRadioGroup
             onChange={updateProps}
-            {...omit(["setProps", "data", "class_name"], props)}
+            {...omit(
+                [
+                    "setProps",
+                    "data",
+                    "class_name",
+                    "persistence",
+                    "persisted_props",
+                    "persistence_type",
+                ],
+                props
+            )}
             className={class_name}
         >
             {data.map((radio, index) => {
@@ -33,8 +43,8 @@ const RadioGroup = (props) => {
 RadioGroup.displayName = "RadioGroup";
 
 RadioGroup.defaultProps = {
-    persisted_props: ['value'],
-    persistence_type: 'local',
+    persisted_props: ["value"],
+    persistence_type: "local",
 };
 
 RadioGroup.propTypes = {
@@ -112,20 +122,20 @@ RadioGroup.propTypes = {
         PropTypes.string,
         PropTypes.number,
     ]),
-                
+
     /**
      * Properties whose user interactions will persist after refreshing the
-     * component or the page. 
+     * component or the page.
      */
-    persisted_props: PropTypes.arrayOf(PropTypes.oneOf(['value'])),
-                
+    persisted_props: PropTypes.arrayOf(PropTypes.oneOf(["value"])),
+
     /**
      * Where persisted user changes will be stored:
      * memory: only kept in memory, reset on page refresh.
      * local: window.localStorage, data is kept after the browser quit.
      * session: window.sessionStorage, data is cleared once the browser quit.
      */
-    persistence_type: PropTypes.oneOf(['local', 'session', 'memory']),
+    persistence_type: PropTypes.oneOf(["local", "session", "memory"]),
 
     /**
      * Adds red asterisk on the right side of label