+++ src/lib/components/Select.react.js
@@ -9,7 +9,17 @@ import { renderDashComponents } from "dash-extensions-js";
  */
 const Select = (props) => {
     const { setProps, data, class_name } = props;
-    let nProps = omit(["setProps", "data", "class_name"], props);
+    let nProps = omit(
+        [
+            "setProps",
+            "data",
+            "class_name",
+            "persistence",
+            "persisted_props",
+            "persistence_type",
+        ],
+        props
+    );
     nProps = renderDashComponents(nProps, [
         "label",
         "description",
@@ -36,8 +46,8 @@ Select.displayName = "Select";
 
 Select.defaultProps = {
     data: [],
-    persisted_props: ['value'],
-    persistence_type: 'local',
+    persisted_props: ["value"],
+    persistence_type: "local",
 };
 
 Select.propTypes = {
@@ -161,7 +171,7 @@ Select.propTypes = {
      * the new `value` also matches what was given originally.
      * Used in conjunction with `persistence_type`.
      */
-     persistence: PropTypes.oneOfType([
+    persistence: PropTypes.oneOfType([
         PropTypes.bool,
         PropTypes.string,
         PropTypes.number,
@@ -169,9 +179,9 @@ Select.propTypes = {
 
     /**
      * Properties whose user interactions will persist after refreshing the
-     * component or the page. 
+     * component or the page.
      */
-    persisted_props: PropTypes.arrayOf(PropTypes.oneOf(['value'])),
+    persisted_props: PropTypes.arrayOf(PropTypes.oneOf(["value"])),
 
     /**
      * Where persisted user changes will be stored:
@@ -179,7 +189,7 @@ Select.propTypes = {
      * local: window.localStorage, data is kept after the browser quit.
      * session: window.sessionStorage, data is cleared once the browser quit.
      */
-    persistence_type: PropTypes.oneOf(['local', 'session', 'memory']),
+    persistence_type: PropTypes.oneOf(["local", "session", "memory"]),
 
     /**
      * Placeholder, displayed when date is not selected