+++ src/lib/components/TimeInput.react.js
@@ -22,7 +22,16 @@ const TimeInput = (props) => {
         setTime(value && new Date(value));
     }, [value]);
 
-    let nProps = omit(["setProps", "class_name"], props);
+    let nProps = omit(
+        [
+            "setProps",
+            "class_name",
+            "persistence",
+            "persisted_props",
+            "persistence_type",
+        ],
+        props
+    );
     nProps = renderDashComponents(nProps, [
         "icon",
         "rightSection",
@@ -44,8 +53,8 @@ const TimeInput = (props) => {
 TimeInput.displayName = "TimeInput";
 
 TimeInput.defaultProps = {
-    persisted_props: ['value'],
-    persistence_type: 'local',
+    persisted_props: ["value"],
+    persistence_type: "local",
 };
 
 TimeInput.propTypes = {
@@ -137,20 +146,20 @@ TimeInput.propTypes = {
         PropTypes.string,
         PropTypes.number,
     ]),
-            
+
     /**
      * Properties whose user interactions will persist after refreshing the
-     * component or the page. 
+     * component or the page.
      */
-    persisted_props: PropTypes.arrayOf(PropTypes.oneOf(['value'])),
-            
+    persisted_props: PropTypes.arrayOf(PropTypes.oneOf(["value"])),
+
     /**
      * Where persisted user changes will be stored:
      * memory: only kept in memory, reset on page refresh.
      * local: window.localStorage, data is kept after the browser quit.
      * session: window.sessionStorage, data is cleared once the browser quit.
      */
-    persistence_type: PropTypes.oneOf(['local', 'session', 'memory']),    
+    persistence_type: PropTypes.oneOf(["local", "session", "memory"]),
 
     /**
      * Input border-radius from theme or number to set border-radius in px